backend/requirements.txt
asgiref==3.2.3
astroid==2.3.3
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
config==0.4.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
Django==3.0.3
django-cors-headers==3.2.1
django-enum-choices==2.1.2
django-rest-swagger==2.2.0
djangorestframework==3.11.0
dnspython==1.16.0
eventlet==0.25.1
greenlet==0.4.15
idna==2.9
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
mysqlclient==1.4.6
openapi-codec==1.3.2 
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
python-dateutil==2.8.1
python-engineio==3.12.1
python-socketio==4.4.0
pytz==2019.3
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
wrapt==1.11.2
XlsxWriter==1.2.7
openpyxl==3.0.3
django-crontab==0.7.1
django-prometheus==2.1.0
django-filter==2.4.0